# api-docs.yml
openapi: 3.0.0
info:
  title: Mockiato
  description: ''
  contact: {}
  version: '1.0.0'
servers:
- url: '{mockiatoUrl}'
  variables:
    mockiatoUrl:
      default: http://mockiato.optum.com/api
paths:
...
  /services:
    post:
      tags:
      - services
      summary: addVirtualService
      description: creates a new virtual service
      operationId: AddVirtualService
      parameters:
      - name: x-access-token
        in: header
        required: true
        description: your access token
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: the service to create
        content:
          application/json:
            schema:
              description: the service to create
              $ref: '#/components/schemas/VirtualService'
        required: false
      responses:
        200:
          description: successful response
          content:
            application/json:
              schema:
                description: successful response
                $ref: '#/components/schemas/VirtualService'
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
...
  /services/{id}:
    get:
      tags:
      - services
      summary: getVirtualServiceByID
      description: retrieves a virtual service
      operationId: GetVirtualServiceByID
      parameters:
      - name: id
        in: path
        description: ID of service to fetch
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful response
          content:
            application/json:
              schema:
                description: successful response
                $ref: '#/components/schemas/VirtualService'
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
    put:
      tags:
      - services
      summary: updateVirtualService
      description: updates a virtual service
      operationId: UpdateVirtualService
      parameters:
      - name: id
        in: path
        description: ID of service to update
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: x-access-token
        in: header
        required: true
        description: your access token
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: the service to update
        content:
          application/json:
            schema:
              description: the service to update
              $ref: '#/components/schemas/VirtualService'
        required: false
      responses:
        200:
          description: successful response
          content:
            application/json:
              schema:
                description: successful response
                $ref: '#/components/schemas/VirtualService'
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
    delete:
      tags:
      - services
      summary: removeVirtualService
      description: removes a virtual service
      operationId: RemoveVirtualService
      parameters:
      - name: id
        in: path
        description: ID of service to remove
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: x-access-token
        in: header
        required: true
        description: your access token
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: successful response
          content:
            application/json:
              schema:
                description: successful response
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
...
components:
  schemas:
...
    VirtualService:
      title: VirtualService
      type: object
      properties:
        sut:
          $ref: '#/components/schemas/SystemUnderTest'
        user:
          $ref: '#/components/schemas/UserDetails'
        name:
          type: string
        type:
          type: string
        basePath:
          type: string
        matchTemplates:
          type: array
          items:
            type: string
        rrpairs:
          type: array
          items:
            $ref: '#/components/schemas/ReqResPair'
          description: ''
        delay:
          type: integer
          format: int64
        running:
          type: boolean
    SystemUnderTest:
      title: SystemUnderTest
      type: object
      properties:
        name:
          type: string
    ReqResPair:
      title: ReqResPair
      type: object
      properties:
        verb:
          type: string
        path:
          type: string
        payloadType:
          type: string
        reqHeaders:
          type: object
        reqData:
          type: object
        resStatus:
          type: integer
          format: int64
        resHeaders:
          type: object
        resData:
          type: object
    UserDetails:
      title: UserDetails
      type: object
      properties:
        uid:
          type: string
        mail:
          type: string
...